Move missing_types core functionality out of parser.
